+++ src/engine/resources/RemoteResourceManager.ts
@@ -1,33 +1,18 @@
-import { GameState } from "../GameWorker";
 import { WorkerMessageType, PostMessageTarget } from "../WorkerMessage";
 import { ResourceDefinition, ResourceState } from "./ResourceManager";
-
-export type RemoteResourceLoaderFactory<RemoteResource = undefined> = (
-  state: GameState
-) => RemoteResourceLoader<RemoteResource>;
-
-export interface RemoteResourceLoader<RemoteResource = undefined> {
-  type: string;
-  load?(resourceId: number, resourceDef: ResourceDefinition): void;
-  loaded?(resourceId: number, remoteResource: RemoteResource): void;
-  loadError?(resourceId: number, error: any): void;
-  addRef?(resourceId: number): void;
-  removeRef?(resourceId: number): void;
-  dispose?(resourceId: number): void;
-}
 
 export interface RemoteResourceManager {
   buffer: SharedArrayBuffer;
   view: Uint32Array;
   postMessageTarget: PostMessageTarget;
   store: Map<number, RemoteResourceInfo<any>>;
-  resourceLoaders: Map<string, RemoteResourceLoader<any>>;
 }
 
 export interface RemoteResourceInfo<RemoteResource = undefined> {
   resourceId: number;
   type: string;
   state: ResourceState;
+  error?: any;
   remoteResource?: RemoteResource;
 }
 
@@ -40,13 +25,7 @@ export function createRemoteResourceManager(
     view: new Uint32Array(buffer),
     postMessageTarget,
     store: new Map(),
-    resourceLoaders: new Map(),
   };
-}
-
-export function registerRemoteResourceLoader(state: GameState, loaderFactory: RemoteResourceLoaderFactory<any>): void {
-  const loader = loaderFactory(state);
-  state.resourceManager.resourceLoaders.set(loader.type, loader);
 }
 
 export function loadRemoteResource<Def extends ResourceDefinition>(
@@ -98,12 +77,6 @@ export function remoteResourceLoaded(manager: RemoteResourceManager, resourceId:
 
   resourceInfo.state = ResourceState.Loaded;
   resourceInfo.remoteResource = remoteResource;
-
-  const loader = manager.resourceLoaders.get(resourceInfo.type)!;
-
-  if (loader.loaded) {
-    loader.loaded(resourceId, remoteResource);
-  }
 }
 
 export function remoteResourceLoadError(manager: RemoteResourceManager, resourceId: number, error: any) {
@@ -114,26 +87,9 @@ export function remoteResourceLoadError(manager: RemoteResourceManager, resource
   }
 
   resourceInfo.state = ResourceState.Error;
-
-  const loader = manager.resourceLoaders.get(resourceInfo.type)!;
-
-  if (loader.loadError) {
-    loader.loadError(resourceId, error);
-  }
+  resourceInfo.error = error;
 }
 
 export function remoteResourceDisposed(manager: RemoteResourceManager, resourceId: number) {
-  const resourceInfo = manager.store.get(resourceId);
-
-  if (!resourceInfo) {
-    return;
-  }
-
-  const loader = manager.resourceLoaders.get(resourceInfo.type)!;
-
-  if (loader.dispose) {
-    loader.dispose(resourceId);
-  }
-
   manager.store.delete(resourceId);
 }
