+++ src/engine/resources/TextureResourceLoader.ts
@@ -1,9 +1,7 @@
 import { Texture } from "three";
 import { RGBELoader } from "three/examples/jsm/loaders/RGBELoader";
 
-import { RemoteResourceManager, loadRemoteResource, RemoteResourceLoader } from "./RemoteResourceManager";
 import { ResourceDefinition, ResourceLoader, ResourceManager } from "./ResourceManager";
-import { GameState } from "../GameWorker";
 
 export const TEXTURE_RESOURCE = "texture";
 
@@ -45,14 +43,4 @@ export function TextureResourceLoader(manager: ResourceManager): ResourceLoader<
       };
     },
   };
-}
-
-export function TextureRemoteResourceLoader(state: GameState): RemoteResourceLoader {
-  return {
-    type: TEXTURE_RESOURCE,
-  };
-}
-
-export function loadRemoteTexture(manager: RemoteResourceManager, textureDef: TextureDef): number {
-  return loadRemoteResource(manager, textureDef);
 }
